Util to handle element not found
We often have to query the dom for an element by selector.
However it may not find anything. To make sure we handle
it correctly and to keep logic in one spot, adding a util
function that shows the error but keeps typescript happy.
If element was not found, simply give it a stub and warn
developer.