fixup! Util to get the list of selected categories
For previously subscribed users with no tags, this change will default all checkboxes to true. Otherwise, the function returns the unchecked categories object which defaults all to unchecked. Brings it inline with behavior of code block below that calls `getCheckedStatusForTagValue` (if tag value is undefined, it returns it as true)