Prompting Refactor: new function to trigger autoPrompts
Motivation:
In `InitHelper` we no longer want to call a single autoPrompt but rather "spawn" all the autoPrompts that are configured. Note: this includes the singular native prompt as well as the multiple possible slidedown prompts (e.g: category, push only, sms, email...)

We also encapsulate the logic as to whether we should force the slidedown even if native prompt is shown into its own function (this is the case in certain browser environments such as when the browser requires a user interaction).