Change to Slidedown EVENT triggering to get test environment to work
Motivation:
This change provides a way for the Slidedown events to be triggered from higher scopes. This is due to the `once` function not running correctly in the test environment. Previously, the CLOSED event was triggered in the callback of the `once` function succeeding. Here, we are moving it one level up to fix unit tests.

Future: figure out a way to get the `once` function to work in test environments.